Use the default SearchQuerySet class without overriding get_queryset
Overriding get_queryset causes problems when using `YourManager.from_queryset(YourQuerySetClass)` instead of return an instances of `YourQuerySetClass`, it will return an instance of `SearchQuerySet`. By changing the implementation of the `PGSearchManager` class, you can support this use case, by not overriding the `get_queryset` method so it will replace the `BaseManager` method, so it can then return the correct `QuerySet` subclass.